This the inital version of Job portal.
It works with some static entries for mongodb. It can be enhanced to take value from GUI in AngularJS as well. the hard coded values are used in middleware function in mymiddleware.js. to show/update/delete the entries for Job/Profile/Company we need to chnage the document id in the collection of mongo db to an actual exiting id( current one will not be valid in the database created again).
BeProdigyJobs.js in the main file to be used as "#node BeProdigyJobs.js". i have used get calls so that it could be easily tested with simple url call from browser/curl.